daemon: split daemon core-loop logic
